backend/requirements.txt
-i https://pypi.org/simple
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.8
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
watchgod==0.8.2
websockets==13.0
